Implement params parsing
Finally.
A little more complicated than expecting due to the necessity
of figuring out the type of the param (which is not the in the
JSON representation for obvious reasons). This is now handled
by the ChannelFactory, which acts as a metaclass for actions/triggers.
I think I wrote too much python in my life.